+++ fabric/src/main/java/com/binaris/wizardry/gametest/ArcaneWorkbenchTest.java
@@ -3,7 +3,7 @@ package com.binaris.wizardry.gametest;
 import com.binaris.wizardry.api.content.DeferredObject;
 import com.binaris.wizardry.content.item.WizardArmorItem;
 import com.binaris.wizardry.content.item.WizardArmorType;
-import com.binaris.wizardry.core.gametest.ArcaneWorkbenchTest;
+import com.binaris.wizardry.core.gametest.AWTestHandler;
 import com.binaris.wizardry.setup.datagen.EBDataGenProcessor;
 import com.binaris.wizardry.setup.registries.EBItems;
 import com.binaris.wizardry.setup.registries.Spells;
@@ -14,31 +14,31 @@ import net.minecraft.world.item.Item;
 import java.util.List;
 
 @SuppressWarnings("unused")
-public class FabricArcaneWorkbenchTest {
+public class ArcaneWorkbenchTest {
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void applySpellsToWand(GameTestHelper helper) {
         EBDataGenProcessor.wandItems().values().forEach(wand ->
-                ArcaneWorkbenchTest.applySpellsToWand(helper, wand.get(), Spells.COBWEBS, Spells.FIREBALL));
+                AWTestHandler.applySpellsToWand(helper, wand.get(), Spells.COBWEBS, Spells.FIREBALL));
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void canUpgradeToNextTier(GameTestHelper helper) {
         EBDataGenProcessor.wandItems().values().forEach(wand ->
-                ArcaneWorkbenchTest.canUpgradeToNextTier(helper, wand.get()));
+                AWTestHandler.canUpgradeToNextTier(helper, wand.get()));
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void putSpellOnBlankScroll(GameTestHelper helper) {
-        ArcaneWorkbenchTest.putSpellOnBlankScroll(helper, Spells.ARCANE_LOCK);
+        AWTestHandler.putSpellOnBlankScroll(helper, Spells.ARCANE_LOCK);
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void putSpellOnScroll(GameTestHelper helper) {
-        ArcaneWorkbenchTest.putSpellOnScrollFilled(helper);
+        AWTestHandler.putSpellOnScrollFilled(helper);
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
@@ -53,7 +53,7 @@ public class FabricArcaneWorkbenchTest {
                 .map(DeferredObject::get)
                 .filter(item -> ((WizardArmorItem) item).getWizardArmorType() == WizardArmorType.WIZARD)
                 .forEach(armor -> upgrades.forEach(upgrade ->
-                        ArcaneWorkbenchTest.upgradeNormalArmor(helper, armor, upgrade)));
+                        AWTestHandler.upgradeNormalArmor(helper, armor, upgrade)));
         helper.succeed();
     }
 
@@ -69,25 +69,25 @@ public class FabricArcaneWorkbenchTest {
                 .map(DeferredObject::get)
                 .filter(item -> ((WizardArmorItem) item).getWizardArmorType() != WizardArmorType.WIZARD)
                 .forEach(armor -> upgrades.forEach(upgrade ->
-                        ArcaneWorkbenchTest.cannotUpgradeMaxedArmor(helper, armor, upgrade)));
+                        AWTestHandler.cannotUpgradeMaxedArmor(helper, armor, upgrade)));
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void repairWand(GameTestHelper helper) {
         EBDataGenProcessor.wandItems().values().forEach(wand ->
-                ArcaneWorkbenchTest.repairWand(helper, wand.get(), EBItems.MAGIC_CRYSTAL.get()));
+                AWTestHandler.repairWand(helper, wand.get(), EBItems.MAGIC_CRYSTAL.get()));
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void cannotExceedBlankScrollLimit(GameTestHelper helper) {
-        ArcaneWorkbenchTest.cannotExceedBlankScrollLimit(helper);
+        AWTestHandler.cannotExceedBlankScrollLimit(helper);
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void cannotExceedSpellBookLimit(GameTestHelper helper) {
-        ArcaneWorkbenchTest.cannotExceedSpellBookLimit(helper, Spells.FIREBALL);
+        AWTestHandler.cannotExceedSpellBookLimit(helper, Spells.FIREBALL);
         helper.succeed();
     }
 
@@ -104,13 +104,13 @@ public class FabricArcaneWorkbenchTest {
         );
 
         upgrades.forEach(upgrade ->
-                ArcaneWorkbenchTest.cannotExceedUpgradeLimit(helper, upgrade));
+                AWTestHandler.cannotExceedUpgradeLimit(helper, upgrade));
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void upgradeWandNextTierNBT(GameTestHelper helper) {
-        EBDataGenProcessor.wandItems().values().forEach(wand -> ArcaneWorkbenchTest.upgradeWandNextTierNBT(helper, wand.get()));
+        EBDataGenProcessor.wandItems().values().forEach(wand -> AWTestHandler.upgradeWandNextTierNBT(helper, wand.get()));
         helper.succeed();
     }
 }