+++ fabric/src/main/java/com/binaris/wizardry/gametest/WandTest.java
@@ -1,44 +1,44 @@
 package com.binaris.wizardry.gametest;
 
-import com.binaris.wizardry.core.gametest.WandTest;
+import com.binaris.wizardry.core.gametest.WandTestHandler;
 import net.minecraft.gametest.framework.GameTest;
 import net.minecraft.gametest.framework.GameTestHelper;
 
 @SuppressWarnings("unused")
-public class FabricWandTest {
+public class WandTest {
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void wandBasicMovement(GameTestHelper helper) {
-        WandTest.wandBasicMovement(helper);
+        WandTestHandler.wandBasicMovement(helper);
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void wandPartiallyEmpty(GameTestHelper helper) {
-        WandTest.wandPartiallyEmpty(helper);
+        WandTestHandler.wandPartiallyEmpty(helper);
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void wandCircularSelection(GameTestHelper helper) {
-        WandTest.wandCircularSelection(helper);
+        WandTestHandler.wandCircularSelection(helper);
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void wandLiteralIndex(GameTestHelper helper) {
-        WandTest.wandLiteralIndex(helper);
+        WandTestHandler.wandLiteralIndex(helper);
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:arcane_workbench_3x3x3")
     public static void wandLiteralIndexPartiallyEmpty(GameTestHelper helper) {
-        WandTest.wandLiteralIndexPartiallyEmpty(helper);
+        WandTestHandler.wandLiteralIndexPartiallyEmpty(helper);
         helper.succeed();
     }
 
     @GameTest(template = "ebwizardry:empty")
     public static void siphonUpgradePlayerKillMob(GameTestHelper helper) {
-        WandTest.siphonUpgradePlayerKillMob(helper);
+        WandTestHandler.siphonUpgradePlayerKillMob(helper);
         helper.succeed();
     }
 }