+++ forge/src/main/java/com/binaris/wizardry/gametest/ArcaneWorkbenchTest.java
@@ -5,7 +5,7 @@ import com.binaris.wizardry.api.EBLogger;
 import com.binaris.wizardry.api.content.DeferredObject;
 import com.binaris.wizardry.content.item.WizardArmorItem;
 import com.binaris.wizardry.content.item.WizardArmorType;
-import com.binaris.wizardry.core.gametest.ArcaneWorkbenchTest;
+import com.binaris.wizardry.core.gametest.AWTestHandler;
 import com.binaris.wizardry.setup.datagen.EBDataGenProcessor;
 import com.binaris.wizardry.setup.registries.EBItems;
 import com.binaris.wizardry.setup.registries.Spells;
@@ -20,12 +20,12 @@ import java.util.List;
 @SuppressWarnings("unused")
 @PrefixGameTestTemplate(false)
 @GameTestHolder(WizardryMainMod.MOD_ID)
-public class ForgeArcaneWorkbenchTest {
+public class ArcaneWorkbenchTest {
 
     @GameTest(template = "arcane_workbench_3x3x3")
     public static void applySpellsToWand(GameTestHelper helper) {
         EBDataGenProcessor.wandItems().values().forEach(wand ->
-                ArcaneWorkbenchTest.applySpellsToWand(helper, wand.get(), Spells.COBWEBS, Spells.FIREBALL));
+                AWTestHandler.applySpellsToWand(helper, wand.get(), Spells.COBWEBS, Spells.FIREBALL));
         helper.succeed();
     }
 
@@ -37,13 +37,13 @@ public class ForgeArcaneWorkbenchTest {
 
     @GameTest(template = "arcane_workbench_3x3x3")
     public static void putSpellOnBlankScroll(GameTestHelper helper) {
-        ArcaneWorkbenchTest.putSpellOnBlankScroll(helper, Spells.ARCANE_LOCK);
+        AWTestHandler.putSpellOnBlankScroll(helper, Spells.ARCANE_LOCK);
         helper.succeed();
     }
 
     @GameTest(template = "arcane_workbench_3x3x3")
     public static void putSpellOnScroll(GameTestHelper helper) {
-        ArcaneWorkbenchTest.putSpellOnScrollFilled(helper);
+        AWTestHandler.putSpellOnScrollFilled(helper);
     }
 
     @GameTest(template = "arcane_workbench_3x3x3")
@@ -58,7 +58,7 @@ public class ForgeArcaneWorkbenchTest {
                 .map(DeferredObject::get)
                 .filter(item -> ((WizardArmorItem) item).getWizardArmorType() == WizardArmorType.WIZARD)
                 .forEach(armor -> upgrades.forEach(upgrade ->
-                        ArcaneWorkbenchTest.upgradeNormalArmor(helper, armor, upgrade)));
+                        AWTestHandler.upgradeNormalArmor(helper, armor, upgrade)));
         helper.succeed();
     }
 
@@ -74,25 +74,25 @@ public class ForgeArcaneWorkbenchTest {
                 .map(DeferredObject::get)
                 .filter(item -> ((WizardArmorItem) item).getWizardArmorType() != WizardArmorType.WIZARD)
                 .forEach(armor -> upgrades.forEach(upgrade ->
-                        ArcaneWorkbenchTest.cannotUpgradeMaxedArmor(helper, armor, upgrade)));
+                        AWTestHandler.cannotUpgradeMaxedArmor(helper, armor, upgrade)));
         helper.succeed();
     }
 
     @GameTest(template = "arcane_workbench_3x3x3")
     public static void repairWand(GameTestHelper helper) {
         EBDataGenProcessor.wandItems().values().forEach(wand ->
-                ArcaneWorkbenchTest.repairWand(helper, wand.get(), EBItems.MAGIC_CRYSTAL.get()));
+                AWTestHandler.repairWand(helper, wand.get(), EBItems.MAGIC_CRYSTAL.get()));
         helper.succeed();
     }
 
     @GameTest(template = "arcane_workbench_3x3x3")
     public static void cannotExceedBlankScrollLimit(GameTestHelper helper) {
-        ArcaneWorkbenchTest.cannotExceedBlankScrollLimit(helper);
+        AWTestHandler.cannotExceedBlankScrollLimit(helper);
     }
 
     @GameTest(template = "arcane_workbench_3x3x3")
     public static void cannotExceedSpellBookLimit(GameTestHelper helper) {
-        ArcaneWorkbenchTest.cannotExceedSpellBookLimit(helper, Spells.FIREBALL);
+        AWTestHandler.cannotExceedSpellBookLimit(helper, Spells.FIREBALL);
         helper.succeed();
     }
 
@@ -109,7 +109,7 @@ public class ForgeArcaneWorkbenchTest {
         );
 
         upgrades.forEach(upgrade ->
-                ArcaneWorkbenchTest.cannotExceedUpgradeLimit(helper, upgrade));
+                AWTestHandler.cannotExceedUpgradeLimit(helper, upgrade));
         helper.succeed();
     }